refactor: extract class generation to different file
